fix: fused SIMD OverflowError + wrong results for long op chains
Critical bugs fixed:

1. OverflowError: out of range integral type conversion
   - FusedOpDesc.lhs_idx and rhs_idx were u8, limiting indices to 0-255
   - With 400+ binops in Mandelbrot, op_dst_to_idx exceeded 255
   - Changed lhs_idx/rhs_idx from u8 to u16 (supports up to 65535)
   - Updated Python bindings: Vec<(u8,u8,u8,u8,u8)> → Vec<(u8,u8,u16,u8,u16)>

2. Silent truncation to 8 ops (wrong results)
   - MAX_FUSED_OPS=8 caused kernel to silently drop ops beyond the 8th
   - Mandelbrot with 700 ops: only 8 executed → mean=-0.5000 instead of -0.7515
   - Scalar fallback now uses Vec<f32/f64> intermediates (arbitrary length)
   - For ops.len() > MAX_FUSED_OPS: scalar path processes ALL ops correctly
   - AVX2 path preserved for ≤8 ops (fast SIMD)

3. Performance: hoisted Vec allocation outside scalar loop
   - Was allocating Vec per element (250K allocations for 500×500 array)
   - Now allocates once, fills with 0.0 per element via .fill(0.0)
   - Same fix applied to both f32 and f64 scalar paths

Files changed:
- rust-engine/src/x86_emitter.rs: FusedOpDesc u8→u16, scalar paths
- rust-engine/src/lib.rs: function signatures u8→u16
- python/src/lib.rs: function signatures u8→u16

+++ python/src/lib.rs
@@ -1506,7 +1506,7 @@ fn simd_reduce_f64(op: &str, data_ptr: usize, n: usize) -> PyResult<f64> {
 /// dst_ptr: output array pointer (used when reduce_op == 255)
 #[pyfunction]
 fn simd_fused_elementwise_f32(
-    ops: Vec<(u8, u8, u8, u8, u8)>,
+    ops: Vec<(u8, u8, u16, u8, u16)>,
     input_ptrs: Vec<usize>,
     constants: Vec<f32>,
     n: usize,
@@ -1520,7 +1520,7 @@ fn simd_fused_elementwise_f32(
 /// Same as simd_fused_elementwise_f32 but for double-precision data.
 #[pyfunction]
 fn simd_fused_elementwise_f64(
-    ops: Vec<(u8, u8, u8, u8, u8)>,
+    ops: Vec<(u8, u8, u16, u8, u16)>,
     input_ptrs: Vec<usize>,
     constants: Vec<f64>,
     n: usize,
